Applied patch [ 1195797 ] Fixes to compile on Win64
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@34241 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/dlmsw.cpp
@@ -80,7 +80,7 @@ public:
     };
 
     static BOOL CALLBACK
-        EnumModulesProc(PSTR name, DWORD base, ULONG size, void *data);
+        EnumModulesProc(PSTR name, DWORD64 base, ULONG size, void *data);
 };
 
 // ----------------------------------------------------------------------------
@@ -209,7 +209,7 @@ wxString wxVersionDLL::GetFileVersion(const wxString& filename) const
 /* static */
 BOOL CALLBACK
 wxDynamicLibraryDetailsCreator::EnumModulesProc(PSTR name,
-                                                DWORD base,
+                                                DWORD64 base,
                                                 ULONG size,
                                                 void *data)
 {
